Cleanup lib/util/ to remove unused code
Move function only used in tests to test files to avoid bundling it
