+++ lib/util/__tests__/utimes.test.js
@@ -11,6 +11,20 @@ let utimes
 
 /* global beforeEach, describe, it */
 
+// HFS, ext{2,3}, FAT do not
+function hasMillisResSync () {
+  let tmpfile = path.join('millis-test-sync' + Date.now().toString() + Math.random().toString().slice(2))
+  tmpfile = path.join(os.tmpdir(), tmpfile)
+
+  // 550 millis past UNIX epoch
+  const d = new Date(1435410243862)
+  fs.writeFileSync(tmpfile, 'https://github.com/jprichardson/node-fs-extra/pull/141')
+  const fd = fs.openSync(tmpfile, 'r+')
+  fs.futimesSync(fd, d, d)
+  fs.closeSync(fd)
+  return fs.statSync(tmpfile).mtime > 1435410243000
+}
+
 describe('utimes', () => {
   let TEST_DIR
 
@@ -20,32 +34,6 @@ describe('utimes', () => {
     // reset stubs
     gracefulFsStub = {}
     utimes = proxyquire('../utimes', { 'graceful-fs': gracefulFsStub })
-  })
-
-  describe('hasMillisResSync()', () => {
-    it('should return a boolean indicating whether it has support', () => {
-      const res = utimes.hasMillisResSync()
-      assert.strictEqual(typeof res, 'boolean')
-
-      // HFS => false
-      if (process.platform === 'darwin') assert.strictEqual(res, false)
-
-      // does anyone use FAT anymore?
-      // if (process.platform === 'win32') assert.strictEqual(res, true)
-      // fails on appveyor... could appveyor be using FAT?
-
-      // this would fail if ext2/ext3
-      if (process.platform === 'linux') assert.strictEqual(res, true)
-    })
-  })
-
-  describe('timeRemoveMills()', () => {
-    it('should remove millisecond precision from a timestamp', () => {
-      const ts = 1334990868773
-      const ets = 1334990868000
-      assert.strictEqual(utimes.timeRemoveMillis(ts), ets)
-      assert.strictEqual(utimes.timeRemoveMillis(new Date(ts)).getTime(), ets)
-    })
   })
 
   describe('utimesMillis()', () => {
@@ -66,7 +54,7 @@ describe('utimes', () => {
       utimes.utimesMillis(tmpFile, awhileAgo, awhileAgo, err => {
         assert.ifError(err)
         stats = fs.statSync(tmpFile)
-        if (utimes.hasMillisResSync()) {
+        if (hasMillisResSync()) {
           assert.deepStrictEqual(stats.mtime, awhileAgo)
           assert.deepStrictEqual(stats.atime, awhileAgo)
         } else {
